% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{print.xs}
\alias{print.xs}
\title{Print an rxspec session}
\usage{
\method{print}{xs}(
  x,
  files = TRUE,
  stat = TRUE,
  fit_params = TRUE,
  model = TRUE,
  chain = TRUE
)
}
\description{
Print an rxspec session
}
