% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{xsession}
\alias{xsession}
\title{Create a new rxspec session}
\usage{
xsession()
}
\value{
An object of class \code{"xs"}.
}
\description{
Creates an R-side XSPEC session object and resets the current PyXspec state.
}
\examples{
\dontrun{
specs <- Sys.glob("~/srg/PG1634+706/data/original/XMM/0852980*01/pps/*min*")

xx <- xsession() |>
  set_parallel_settings(
    leven = 20L,
    error = 20L,
    walkers = 20L
  ) |>
  set_fit_settings(
    nIterations = 100L,
    statMethod = "cstat",
    query = "yes"
  ) |>
  load_data(
    specs = specs,
    grp_dat = seq_along(specs),
    grp_plt = seq_along(specs)
  ) |>
  ignore("bad") |>
  ignore_for_all("**-0.5 10.-**") |>
  ignore("1-3: **-0.5 10.0-**") |>
  make_model("const*TBabs*pow") |>
  set_params(2L, "0.0574 -1") |>
  set_params_by_name(
    cflux.Emin = "0.5",
    cflux.Emax = 3.0,
    zgauss_7.LineE = 30
  ) |>
  untie_params(seq(1, 36, 4)) |>
  link_params(5 ~ 1, 9 ~ 0.99 * 1) |>
  link_params(17 ~ 13, 21 ~ 13) |>
  link_params(33 ~ 25, 29 ~ 25) |>
  fit_model()

print(xx)
}
}
